pyproject.toml
[project]
name = "confluence-mcp-server"
version = "0.1.0"
description = "MCP server for Confluence search"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=0.1.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[project.scripts]
confluence-mcp = "confluence_mcp.server:mcp.run"
